// src/src/features/Home/types.ts
export interface ITopNMostSoldProduct {
    productCode: string;
    productName: string;
    soldQuantity: number;
    soldPrice: number;
    productAvailableQuantity: number;
    mainGroupName: string;
    sideGroupName: string;
    formattedSoldPrice: string;
    soldPriceUOM: string;
    id: string;
}

export interface ITopSellingProducts {
    topNMostProductsByPrices: ITopNMostSoldProduct[];
    topNMostProductsByQuantity: ITopNMostSoldProduct[];
}

export interface ITopNMostRevenuableProduct {
    productName: string;
    productCode: string;
    salesQuantity: number;
    salesRevenuAmount: number;
    formattedSalesRevenuAmount: string;
    salesRevenuAmountUOM: string;
    revenuPercentage: number;
    purchaseAmount: number;
    formattedPurchaseAmount: string;
    purchaseAmountUOM: string;
    saleAmount: number;
    formattedSaleAmount: string;
    saleAmountUOM: string;
}